requirements.txt
absl-py==2.1.0
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.7
decorator==5.1.1
distlib==0.3.9
distro==1.9.0
duckduckgo_search==6.3.3
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.6.0
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.12
langchain-google-genai==2.0.1
langchain-openai==0.2.3
langchain-text-splitters==0.3.0
langsmith==0.1.137
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numexpr==2.10.1
numpy==1.26.4
openai==1.52.1
opt_einsum==3.4.0
optree==0.13.0
orjson==3.10.10
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pipenv==2024.2.0
platformdirs==4.3.6
primp==0.6.5
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.3
rsa==4.9
safetensors==0.4.5
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.16.0
tiktoken==0.8.0
tokenizers==0.20.1
torch==2.2.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.27.0
wcwidth==0.2.13
Werkzeug==3.0.4
wikipedia==1.4.0
wrapt==1.16.0
yarl==1.16.0
